refactor(api): simplify subscriber types/entities/interfaces
