+++ libs/dal/src/repositories/subscriber/subscriber.entity.ts
@@ -1,10 +1,10 @@
-import { ChatProviderIdEnum, PushProviderIdEnum, SubscriberCustomData } from '@novu/shared';
+import { IChannelSettings, ISubscriber, SubscriberCustomData } from '@novu/shared';
 import type { ChangePropsValueType } from '../../types/helpers';
 import type { EnvironmentId } from '../environment';
 import type { OrganizationId } from '../organization';
 import { ExternalSubscriberId } from './types';
 
-export class SubscriberEntity {
+export class SubscriberEntity implements ISubscriber {
   // TODO: Use SubscriberId. Means lot of changes across whole codebase. Cool down.
   _id: string;
 
@@ -22,6 +22,7 @@ export class SubscriberEntity {
 
   subscriberId: ExternalSubscriberId;
 
+  // @deprecated: use channelEndpoints instead
   channels?: IChannelSettings[];
 
   topics?: string[];
@@ -48,18 +49,3 @@ export class SubscriberEntity {
 }
 
 export type SubscriberDBModel = ChangePropsValueType<SubscriberEntity, '_environmentId' | '_organizationId'>;
-
-export class IChannelSettings {
-  _integrationId: string;
-
-  providerId: ChatProviderIdEnum | PushProviderIdEnum;
-
-  credentials: IChannelCredentials;
-}
-
-export class IChannelCredentials {
-  phoneNumber?: string;
-  webhookUrl?: string;
-  channel?: string;
-  deviceTokens?: string[];
-}
