--- packages/shared/src/entities/subscriber/index.ts
@@ -1,1 +1,0 @@
-export * from './subscriber.interface';
